test_system/requirements.txt
django
future
djangorestframework
markdown
django-filter
fulltext
numpy
xlrd
xlsxwriter
psycopg2
